+++ lib/CredentialIssuancePurpose.js
@@ -1,16 +1,17 @@
 /*!
- * Copyright (c) 2019-2021 Digital Bazaar, Inc. All rights reserved.
+ * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
-'use strict';
-const jsonld = require('jsonld');
-const {AssertionProofPurpose} = require('jsonld-signatures').purposes;
+import jsonld from 'jsonld';
+import jsigs from 'jsonld-signatures';
+
+const {purposes: {AssertionProofPurpose}} = jsigs;
 
 /**
  * Creates a proof purpose that will validate whether or not the verification
  * method in a proof was authorized by its declared controller for the
  * proof's purpose.
  */
-class CredentialIssuancePurpose extends AssertionProofPurpose {
+export class CredentialIssuancePurpose extends AssertionProofPurpose {
   /**
    * @param {object} options - The options to use.
    * @param {object} [options.controller] - The description of the controller,
@@ -77,5 +78,3 @@ class CredentialIssuancePurpose extends AssertionProofPurpose {
     }
   }
 }
-
-module.exports = CredentialIssuancePurpose;
