+++ lib/contexts/index.js
@@ -1,17 +1,23 @@
 /*!
- * Copyright (c) 2019-2021 Digital Bazaar, Inc. All rights reserved.
+ * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
-'use strict';
+import {
+  contexts as credentialContexts
+} from 'credentials-context';
+import {
+  CONTEXT as vcExamplesV1Context,
+  CONTEXT_URL as vcExamplesV1ContextUrl
+} from './vc-examples-v1.js';
+import {
+  CONTEXT as odrlContext,
+  CONTEXT_URL as odrlContextUrl
+} from './odrl.js';
 
-const {contexts, constants: contextConstants} = require('credentials-context');
+export const contexts = {};
 
-const exportedContexts = module.exports = {
-  'https://www.w3.org/2018/credentials/examples/v1':
-    require('./vc-examples-v1'),
-  'https://www.w3.org/ns/odrl.jsonld': require('./odrl')
-};
+contexts[vcExamplesV1ContextUrl] = vcExamplesV1Context;
+contexts[odrlContextUrl] = odrlContext;
 
-for(const c in contextConstants) {
-  const contextUrl = contextConstants[c];
-  exportedContexts[contextUrl] = contexts.get(contextUrl);
+for(const [url, context] of credentialContexts.entries()) {
+  contexts[url] = context;
 }
