+++ lib/contexts/vc-examples-v1.js
@@ -1,9 +1,10 @@
 /*!
- * Copyright (c) 2019-2021 Digital Bazaar, Inc. All rights reserved.
+ * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
 
+export const CONTEXT_URL = 'https://www.w3.org/2018/credentials/examples/v1';
 /* eslint-disable quote-props */
-module.exports = {
+export const CONTEXT = {
   '@context': [
     {
       '@version': 1.1
