+++ lib/documentLoader.js
@@ -1,12 +1,10 @@
 /*!
- * Copyright (c) 2019-2021 Digital Bazaar, Inc. All rights reserved.
+ * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
-'use strict';
-
 // load locally embedded contexts
-const contexts = require('./contexts');
+import {contexts} from './contexts/index.js';
 
-module.exports = async function documentLoader(url) {
+export async function documentLoader(url) {
   const context = contexts[url];
   if(context !== undefined) {
     return {
@@ -16,4 +14,4 @@ module.exports = async function documentLoader(url) {
     };
   }
   throw new Error(`Document loader unable to load URL "${url}".`);
-};
+}
