--- test/.eslintrc
@@ -1,8 +1,0 @@
-{
-  "env": {
-    "mocha": true
-  },
-  globals: {
-    "should": true
-  }
-}
