+++ test/.eslintrc.cjs
@@ -1,0 +1,8 @@
+module.exports = {
+  env: {
+    mocha: true
+  },
+  globals: {
+    should: true
+  }
+};
