+++ test/constants.js
@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   CREDENTIALS_CONTEXT_URL: 'https://www.w3.org/2018/credentials/v1',
   DID_CONTEXT_URL: 'https://www.w3.org/ns/did/v1',
   VERES_ONE_CONTEXT_URL: 'https://w3id.org/veres-one/v1'
